refactor: KAN-403 알림 로직 리팩토링 및 redis stream 적용 (#84)
* feat: KAN-404 브로드캐스트 알림 읽음 처리 로직 고도화

* feat: notification_read_status 테이블 DDL 추가

* test: KAN-405 브로드캐스트 알림 읽음 처리 테스트코드 작성

* chore: @EnableScheduling 설정을 SchedulingConfig로 분리하여 공통화

* refactor: KAN-403 Redis Pub/Sub 구조를 Redis Stream으로 리팩토링

* refactor: emitter 최대 연결 수 제어 로직 추가

* refactor: 피드 등록 알림 전송 배치 처리, redis stream 발행 로직으로 변경

* refactor: createdAt 정렬 로직 추가 (사용x, redis stream에서 통합 관리)

* feat: sse, redis stream 관련 설정값 추가

+++ src/main/java/com/team1/otvoo/notification/service/NotificationServiceImpl.java
@@ -5,16 +5,22 @@ import com.team1.otvoo.exception.RestException;
 import com.team1.otvoo.notification.dto.NotificationDto;
 import com.team1.otvoo.notification.dto.NotificationDtoCursorResponse;
 import com.team1.otvoo.notification.entity.Notification;
+import com.team1.otvoo.notification.entity.NotificationReadStatus;
 import com.team1.otvoo.notification.mapper.NotificationMapper;
+import com.team1.otvoo.notification.repository.NotificationReadStatusRepository;
 import com.team1.otvoo.notification.repository.NotificationRepository;
+import com.team1.otvoo.user.entity.User;
+import com.team1.otvoo.user.repository.UserRepository;
 import java.time.Instant;
 import java.time.format.DateTimeParseException;
+import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,11 +30,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class NotificationServiceImpl implements NotificationService {
 
   private final NotificationRepository notificationRepository;
+  private final UserRepository userRepository;
+  private final NotificationReadStatusRepository readStatusRepository;
   private final NotificationMapper notificationMapper;
 
   @Transactional(readOnly = true)
   @Override
-  public NotificationDtoCursorResponse getList(UUID receiverId, String cursor, UUID idAfter, int limit) {
+  public NotificationDtoCursorResponse getList(UUID receiverId, String cursor, UUID idAfter,
+      int limit) {
 
     Instant cursorInstant = null;
     if (cursor != null) {
@@ -40,7 +49,8 @@ public class NotificationServiceImpl implements NotificationService {
     }
 
     // hasNext 판별 위해 limit+1 개로 조회
-    List<Notification> notifications = notificationRepository.findNotificationsWithCursor(receiverId, cursorInstant, idAfter, limit+1);
+    List<Notification> notifications = notificationRepository.findUnreadNotificationsWithCursor(
+        receiverId, cursorInstant, idAfter, limit + 1);
 
     boolean hasNext = notifications.size() > limit;
     List<Notification> notificationList = hasNext ? notifications.subList(0, limit) : notifications;
@@ -57,7 +67,7 @@ public class NotificationServiceImpl implements NotificationService {
     }
 
     List<NotificationDto> data = notificationMapper.toDtoList(notificationList);
-    long totalCount = notificationRepository.countByReceiverId(receiverId);
+    long totalCount = notificationRepository.countUnreadNotifications(receiverId);
 
     Notification lastNotification = notificationList.get(notificationList.size() - 1);
     String nextCursor = lastNotification.getCreatedAt().toString();
@@ -74,10 +84,53 @@ public class NotificationServiceImpl implements NotificationService {
 
   @Transactional
   @Override
-  public void delete(UUID notificationId) {
+  public void readNotification(UUID notificationId, UUID userId) {
     Notification notification = notificationRepository.findById(notificationId)
         .orElseThrow(() -> new RestException(ErrorCode.NOT_FOUND, Map.of("id", notificationId)));
 
-    notificationRepository.delete(notification);
+    User receiver = notification.getReceiver();
+    // 개인 알림인 경우
+    if (receiver != null) {
+      // 본인의 알림이 맞는지 확인
+      if (!receiver.getId().equals(userId)) {
+        throw new RestException(ErrorCode.ACCESS_DENIED, Map.of("id", receiver.getId()));
+      }
+      notificationRepository.delete(notification);
+    }
+    // 브로드캐스트 알림인 경우
+    else {
+      // 이미 읽음 처리되었는지 확인
+      boolean alreadyRead = readStatusRepository.existsByUserIdAndNotificationId(userId,
+          notificationId);
+      if (alreadyRead) {
+        log.debug("이미 읽음 처리된 브로드캐스트 알림입니다. notificationId={}, userId={}", notificationId, userId);
+        return;
+      }
+
+      User user = userRepository.findById(userId)
+          .orElseThrow(() -> new RestException(ErrorCode.NOT_FOUND, Map.of("userId", userId)));
+
+      NotificationReadStatus readStatus = new NotificationReadStatus(user, notification);
+      readStatusRepository.save(readStatus);
+    }
   }
+
+  @Transactional
+  @Scheduled(cron = "0 0 4 * * *")
+  @Override
+  public void cleanupOldNotifications() {
+    log.info("오래된 브로드캐스트 알림 데이터 정리 시작");
+    Instant cutoffDateTime = Instant.now().minus(7, ChronoUnit.DAYS);
+    try {
+      long deletedStatusCount = readStatusRepository.deleteByCreatedAtBefore(cutoffDateTime);
+      log.info("Deleted {} notification read statuses older than 7 days", deletedStatusCount);
+
+      long deletedNotificationCount = notificationRepository.deleteByReceiverIsNullAndCreatedAtBefore(
+          cutoffDateTime);
+      log.info("Deleted {} broadcast notifications older than 7 days", deletedNotificationCount);
+    } catch (Exception e) {
+      log.error("Error while deleting old notifications", e);
+    }
+  }
+
 }
